Encapsulate exceptions during chain and chainError
If the function passed to `chain` or `chainError` returned a non-promise
type (specifically, a type that does not implement `map` or
`onRejected`) then an exception would be thrown. This would typically
happen if an exception was thrown during the passed function. Now use
any exception thrown during a `chain` or `chainError` to reject the
resulting promise.